backend/requirements.txt
Flask
Flask-Cors
Flask-Limiter
python-dotenv
requests
authlib
pymongo